updated write in .txt file CLI, now fully functional
